safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.1_Block/S12.1_A4_T2-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected token '}'. Expected a ':' following the property name 'x'.
S12.1_A4_T2

PASS Expected parsing failure

TEST COMPLETE
